refactor: import interface
